# unix: remove busy loop from uv_async_send