The current fix (libuv#2231) was found to be slow in certain cases. This change should improve scalabaility a bit by only incurring the spin loop delay while closing an UV_ASYNC. It also is intended to slightly improve the behavior after uv_loop_close is called, by parking all of the pending flags as set, so that it will not access the loop at all (until the uv_async_t memory is freed, which we leave still to the responsibility of the user).

Takes some inspiration from https://github.com/libuv/libuv/pull/2654
Takes some inspiration from https://github.com/libuv/libuv/pull/2656
Refs: https://github.com/libuv/libuv/pull/2231