@vtjnash 2023-01-14
The syscall expense should be much higher already (since it likely needs to take at least a couple locks), so I am not too worried about that. Later, I think seqcst may be a good promise to add to this function, since it is already mandatory for the syscall code-path to run correctly, but currently the early exit does not guarantee that for all calls. We have run into bugs in Julia that showed we must do a full seqcst fence before this function to use it correctly on non-x86 architectures, which is slightly more expensive.

I hadn't checked cache lines for these, but putting them together seems to make sense, and perhaps even better that we can reuse/rename a `reserved` qword for this.