@vtjnash 2023-01-17
> fd and pending are ~72 bytes apart so two cache lines are locked now instead of one. Maybe use reserved[1] instead?

I looked closer into this, but fd and reserved[1] appear to be the same cache line (and &fd === &reserved[0])

Are you okay with this anyways? It should perform much better than the current spinlock, and v2 can reorder the fields to avoid the issue.

It is also possible to bitpack this all into the current `pending` flag, but possibly a bit more confusing code. Here is the way to do that:
```
atomic_fetch_add(&pending, 2); // increment busy counter
if (atomic_fetch_or(&pending, 1) & 1 == 0) // set pending bit
    write_eventfd()
atomic_fetch_add(&pending, -2); // decrement busy counter
```

Or can even merge the first 2 atomics into 1, but performance should be about the same for this as the above:
```
current = atomic_load_relaxed(&pending);
if (current & 1)
    return;
while (!atomic_compare_exchange_weak(&pending, &current, current + 3))
    if (current & 1)
        return;
write_eventfd();
atomic_fetch_add_relaxed(&pending, -2); // decrement busy counter
```

This might be advantageous because we only need acq-rel instead of seq-cst for internal correctness (though user programs might be relying on the seq-cst barrier, so I would hesitate to remove that)