@bnoordhuis 2023-01-18
> Are you okay with this anyways?

Yes, I am. There are 9 `uv_async_t` benchmarks; it would be interesting to see before/after numbers.

I can try that later this week if you don't get round to it.