@vtjnash 2023-01-18
I don't see any consistent trends (the results below are not exactly repeatable, and I didn't run this on my customized benchmarking machine).

before
```
ok 1 - async1                                                                               
# async1: 9.46 sec (105,718/sec)                                                                                                                                                         
ok 2 - async2                                                                               
# async2: 7.85 sec (127,444/sec)                                                            
ok 3 - async4                                                                                                                                                                            
# async4: 12.67 sec (78,909/sec)                                                            
ok 4 - async8                                                                               
# async8: 25.73 sec (38,867/sec)                                                                                                                                                         
ok 5 - async_pummel_1                                                                       
# async_pummel_1: 1,000,000 callbacks in 5.60 seconds (178,707/sec)
ok 6 - async_pummel_2                                                                       
# async_pummel_2: 1,000,000 callbacks in 5.52 seconds (181,073/sec)
ok 7 - async_pummel_4                                                                                                                                                                    
# async_pummel_4: 1,000,000 callbacks in 6.68 seconds (149,790/sec)
ok 8 - async_pummel_8                
# async_pummel_8: 1,000,000 callbacks in 5.34 seconds (187,129/sec)
ok 13 - million_async
# 6,366,005 async events in 5.0 seconds (1,273,201/s, 1,048,576 unique handles seen)
ok 23 - queue_work
# 215,264 async jobs in 5.0 seconds (43,052/s)
```
after
```
ok 1 - async1
# async1: 10.49 sec (95,337/sec)
ok 2 - async2
# async2: 11.45 sec (87,348/sec)
ok 3 - async4
# async4: 12.28 sec (81,407/sec)
ok 4 - async8
# async8: 25.21 sec (39,659/sec)
ok 5 - async_pummel_1
# async_pummel_1: 1,000,000 callbacks in 5.45 seconds (183,577/sec)
ok 6 - async_pummel_2
# async_pummel_2: 1,000,000 callbacks in 5.79 seconds (172,588/sec)
ok 7 - async_pummel_4
# async_pummel_4: 1,000,000 callbacks in 6.87 seconds (145,522/sec)
ok 8 - async_pummel_8
# async_pummel_8: 1,000,000 callbacks in 7.07 seconds (141,445/sec)
ok 13 - million_async
# 6,249,787 async events in 5.0 seconds (1,249,957/s, 1,048,576 unique handles seen)
ok 1 - queue_work
# 236,608 async jobs in 5.0 seconds (47,321/s)
```