@vtjnash 2023-01-18
I ran ` perf record --call-graph dwarf` on linux, and only 1% of the benchmark time is spent in this function itself
<img width="659" alt="image" src="https://user-images.githubusercontent.com/330950/213257432-1cfbda31-8281-4392-a28b-dd93f821273f.png">